+++ instrumentation/opentelemetry-instrumentation-flask/src/opentelemetry/instrumentation/flask/__init__.py
@@ -266,6 +266,7 @@ from opentelemetry.instrumentation.propagators import (
 )
 from opentelemetry.instrumentation.utils import _start_internal_or_server_span
 from opentelemetry.metrics import get_meter
+from opentelemetry.semconv.attributes.http_attributes import HTTP_ROUTE
 from opentelemetry.semconv.metrics import MetricInstruments
 from opentelemetry.semconv.metrics.http_metrics import (
     HTTP_SERVER_REQUEST_DURATION,
@@ -340,12 +341,16 @@ def _rewrapped_app(
         )
 
         active_requests_counter.add(1, active_requests_count_attrs)
+        request_route = None
 
         def _start_response(status, response_headers, *args, **kwargs):
             if flask.request and (
                 excluded_urls is None
                 or not excluded_urls.url_disabled(flask.request.url)
             ):
+                nonlocal request_route
+                request_route = flask.request.url_rule
+
                 span = flask.request.environ.get(_ENVIRON_SPAN_KEY)
 
                 propagator = get_global_response_propagator()
@@ -388,6 +393,12 @@ def _rewrapped_app(
             duration_attrs_old = otel_wsgi._parse_duration_attrs(
                 attributes, _HTTPStabilityMode.DEFAULT
             )
+
+            if request_route:
+                duration_attrs_old[SpanAttributes.HTTP_TARGET] = str(
+                    request_route
+                )
+
             duration_histogram_old.record(
                 max(round(duration_s * 1000), 0), duration_attrs_old
             )
@@ -395,6 +406,10 @@ def _rewrapped_app(
             duration_attrs_new = otel_wsgi._parse_duration_attrs(
                 attributes, _HTTPStabilityMode.HTTP
             )
+
+            if request_route:
+                duration_attrs_new[HTTP_ROUTE] = str(request_route)
+
             duration_histogram_new.record(
                 max(duration_s, 0), duration_attrs_new
             )